Made changes on Domain and Service

+++ blackboard2/src/main/java/com/example/demo/Dao/AttendanceDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.Domain.Attendance;
 
-public interface AttendanceDao extends JpaRepository<Attendance, Integer> {
+public interface AttendanceDao extends JpaRepository<Attendance,Long > {
 
 }
