+++ blackboard2/src/main/java/com/example/demo/Service/Student_courseService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.example.demo.Domain.Student_course;
 
 public interface Student_courseService {
-	Student_course findone(Integer stcsId);
+	Student_course findone(long stcsId);
 	Student_course  create(Student_course student_course);
 	List<Student_course> findAll();
 }
